+++ db/migrate/20151106141707_add_quote_preference_to_user.rb
@@ -1,0 +1,5 @@
+class AddQuotePreferenceToUser < ActiveRecord::Migration
+  def change
+    add_column :users, :include_quote_in_reply, :boolean, default: true, null: false
+  end
+end